standout in prefs the spellcheck feature fix #11811 - this makes spellcheck a standalone list in the UI to avoid bad UX also moves spellcheck under advanced tab per Brad request Auditors: @darkdh
