+++ app/extensions/brave/locales/en-US/preferences.properties
@@ -117,6 +117,7 @@ downloadsAskAlwaysSwitch=Always ask me where to save files
 downloadsInput=~/downloads/
 downloadsLabel=Save my downloads here:
 emailAddress=Email address
+spellcheck=Spell Check Options
 enableSpellCheck=Enable Spell Check *
 enableAutofill=Enable Autofill
 enableFlash=Enable Adobe Flash support
